requirements.txt
# trend-radar-tw requirements
yfinance>=0.2.36
pandas>=2.1.0
numpy>=1.24.0
pyarrow>=14.0.0
lxml>=5.0.0
html5lib>=1.1
beautifulsoup4>=4.12.0
requests>=2.31.0
pytrends>=4.9.0
anthropic>=0.39.0
streamlit>=1.30.0
plotly>=5.18.0
python-dotenv>=1.0.0
